Implement mandatory validation for task name

+++ app/src/main/java/com/example/retime/CalendarTask.java
@@ -187,13 +187,17 @@ public class CalendarTask extends AppCompatActivity {
             String secondtask = secondtasket.getText().toString();
             String secondtime = secondtimeet.getText().toString();
             String todaydate = _dayofmonth.concat(_month).concat(_year);
-            insertdata(todaydate, firsttask, firsttime, secondtask, secondtime);
-            if (TextUtils.isEmpty(firsttask) && TextUtils.isEmpty(firsttime) && TextUtils.isEmpty(secondtask) && TextUtils.isEmpty(secondtime)){
-                deletebtn.setVisibility(View.INVISIBLE);
-                long id = db.delete(CalendarDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = " + todaydate, null);
-            }
-            else {
-                deletebtn.setVisibility(View.VISIBLE);
+            if ((TextUtils.isEmpty(firsttask) && !TextUtils.isEmpty(firsttime)) || TextUtils.isEmpty(secondtask) && !TextUtils.isEmpty(secondtime)) {
+                Toast.makeText(CalendarTask.this, "Please enter task name", Toast.LENGTH_LONG).show();
+            } else {
+
+                insertdata(todaydate, firsttask, firsttime, secondtask, secondtime);
+                if (TextUtils.isEmpty(firsttask) && TextUtils.isEmpty(firsttime) && TextUtils.isEmpty(secondtask) && TextUtils.isEmpty(secondtime)) {
+                    deletebtn.setVisibility(View.INVISIBLE);
+                    long id = db.delete(CalendarDatabase.TABLE_NAME, CalendarDatabase.COL_1 + " = " + todaydate, null);
+                } else {
+                    deletebtn.setVisibility(View.VISIBLE);
+                }
             }
         }
     }
